requirements.txt
ccxt>=4.0
pandas>=2.0
numpy>=1.24
ta>=0.10
schedule>=1.2
python-dotenv>=1.0
